// Property-Rental-UI/src/app/Models/Property.model.ts
export interface Property {
  propertiesId: number;
  ownerId: number;
  status: string;
  propertyName: string;
  propertyEircode: string;
  rentalValue: number;
  propertyType: string;
  propertyDescription: string;
  bedrooms: number;
  bathrooms: number;
  parkingSpaces: number;
  features: string;
}